yahoo!

+++ src/utils.h
@@ -11,7 +11,7 @@
 namespace sweepmap {
 
 using hash_t     = uint64_t;
-using rpos_t     = int64_t;  // reference position
+using rpos_t     = int32_t;  // reference position; klib anyway doesn't support 64-bit
 using qpos_t     = int32_t;  // query position
 using segm_t     = int32_t;
